+++ data/plugins/generic/polylang/v1/config-plugin.yml
@@ -1,0 +1,2 @@
+src: web
+activate: yes
