+++ src/wordpress/plugins/polylang.py
@@ -1,0 +1,3 @@
+
+class WPPolylang(WPPluginConfig):
+    pass
